refactor get all tasks on date method

+++ Pomodoro.Api/ViewModels/CompletedViewModel.cs
@@ -29,7 +29,7 @@ namespace Pomodoro.Api.ViewModels
         /// <summary>
         /// Gets or sets a value of the spent dime of the pomodoro.
         /// </summary>
-        [Range(1, byte.MaxValue, ErrorMessage = "The {0} property must be in the range from {1} to {2}.")]
+        [Range(1, int.MaxValue, ErrorMessage = "The {0} property must be in the range from {1} to {2}.")]
         public int TimeSpent { get; set; }
 
         /// <summary>
